rename: 관리자 view, 사용자 view 분리
