+++ src/main/java/roomescape/controller/api/TimeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import java.net.URI;
 import java.time.LocalDate;
